Add Hoare specs and quadruple desugaring
All theorems will be stated as doubles desugared from full quadruples
with ghost state. As a sanity check, we have provide a proof that
stating correctness in terms of doubles and specs are the same.

The correctness equivalence proof exposes a small difference between
specs: except for the crash invariant - the double-based definition
allows the proof to skip proving that the postcondition implies the
crash invariant, using the assumption about the continuation rx instead.
As long as the specification satisfies postcond => crashinv, both
program correctness statements are equivalent.