Prove postcondition for RD operations
The Write specifications are far too complicated at the moment. There
are a large number of cases, and perhaps handling them separately rather
than trying to express specifications that cover many possibilities,
would lead to overall shorter proofs with more automation and small,
focused lemmas.

The fact that crash invariants have to re-prove the postcondition is
also really painful, when it would be easy to prove the spec satisfies
postcond => crashinv.