Factor out Write_ok as well
Reveals that the crash refinement doesn't quite go through, due to a
need for recovery in the case of a crash during Write while both disks
are still working. The recovery procedure actually only needs to find a
single address that doesn't match, but this requires a search over the
whole domain of the disk.