Make disk a sized, contiguous memory
Worked out surprisingly well, other than needing more accessor
functions.  Ironically the [D.State] record is still there and didn't
help make this change, so we only have more record accessors.

One annoyance remaining is that for disks it's now natural to test if
the address is in the domain by checking if it's less than the size, but
everywhere else we use the more general check on the disk memory itself.
Checking the memory is probably better but a bit awkward.

Currently we explicitly call [diskMem], which can be used for printing
but is not a disk >-> Funclass coercion so it can't be used to do [d a]
and access an address within a disk. We should probably use a coercion
to Funclass instead.

The code still doesn't know how big the disk is, but we could have a
parameter and assume it's correct in the theorem, or even add an
operation to get the disk size.