Adjust the definition of pred_except
Allows to prove that in [pred_except (mem_is m)] the value can be
swapped out if the memory has a new value.

Note that pred_except is more precise than just removing [a] from the
domain of F: it asserts that F holds when the memory has a |-> v, even
if m already has a |-> v for example.