Prove a (weak) recovery interpretation proof
Currently makes assumptions that are too strong about how the recovery
procedure works (everything requires the invariant and that the
abstraction is preserved, which isn't even true for the replicated
disk). Nonetheless the structure is present even with this proof, and
there are TODOs marking where I think crash invariant handling should be
added.