Progress towards refinement of rexec
Intuitively should be able to take the interpretation refinement
statement and apply it to an interpreted spec-level recovery program, to
prove a stronger statement about recovery execution refinement based
only on this crash refinement property. Seems this is slightly tricky,
though, and requires a few more careful lemmas, or some side condition
is missing.