Re-do replicated disk without separation logic
Crucial simplification came from using md_pred to express a predicate
over an optional disk. Stating a property for when the disk is not
present is not useful, since the disk could be working (allowing any
prop) and then fail in the middle - the important property to capture is
that predicates over the disk are stable.

Another important point is to maintain use of these predicate specs as
long as possible, including in the specs for recovery and as part of the
recovery proofs. Then conversion to the invariant can be done exactly
once. The conversion is also vastly easier, with no difficulties arising
from OOB addresses, when using simple predicates. The only ghost state
is a disk both disks are equal to, if they exist, which should easily be
instantiated to be the abstraction.

Some of the proofs are removed by this commit but will be re-added with
the new, simpler infrastructure.