Prove some more specs for fixup
Might be simpler to have fixup always read the whole disk. Right now the
invariant is that either fixup has nothing left to do, or it has one
address and its position is bounded; a simpler invariant is that the
addresses > a are all handled, which fixup extends by one address. Then
an appropriate functional extensionality (with a separate fact about
equal domains) gives equality of the disks.