Provide a stronger guarantee of recovery progress
Postcondition is now conditional on precondition: in the case where the
disks are equal, recovery (and even lowest-level spec, fixup_ok) promise
to preserve those disks. The ghost state also now makes sense: there is
no address and block if the disk is fully synced.

The one downside is that the automation can't handle instantiating this
ghost state, though it is possible to automate: of one of the two cases
can be solved (with evars for the address and block in the out-of-sync
case), then that's how the ghost state should be instantiated.