Also keep track of Haskell comments
This code has much further to go to being readable. It also arguably
doesn't all have to be uniformly good - students will need to read
little to none of it for the labs. Nonetheless the high-level structure,
and how Coq fits in, is probably interesting for some people, and worth
explaining.