Remove step parameter from all specs
[step] is a global axiom that avoids reasoning directly about what the
axiomatic primitives inside IO do, which in turn prevent prog from just
being Rets and Binds, which provably don't modify the world state.
There's no need for a semantics of operations over abstract states (and
indeed the parameter was never used, just passed around).

This issued stemmed from fixing Semantics so it is unrelated to the
underlying operations - a Semantics is really just a transition
relation, with a known state and return type, and the global [step] is a
semantics for every (again, axiomatic) primitive operation.