Use an interface spec to prove a clean spec
Rewrite TwoDiskProgTheorems.v to re-spec any TwoDiskAPI module.

Some usage of the interface spec, in a manner similar to how we went
from the semantics to a spec with predicates and useful ghost state
(even though for every state, there's a ghost state that makes the
precondition true).

Rather than tying the semantics to [Prim op] as we did before, we will
now tie them to full [prog]s, albeit axiomatic since this is the lowest
level.