Prove rok to rspec using a constraint on rec
Require that rec do nothing when run with the invariant. This is
stronger than strictly necessary - the recovery procedure merely has to
preserve the postcondition of each operation, which might not be
captured by the invariant. However, this is a simple property of
recovery that can be proven once and re-used, and the recovery
procedures I can think of so far should satisfy it.