Attempt to support rok proofs ignoring finish case
Before, we could prove _ok theorems that had crash invariants which did
not include the postcondition (leaving it to the continuation). It's not
so clear how to do this for recovery, where going from p's Recovered
execution to one for rx to the same state seems to require something
about recovering twice in a row.

The consequence is merely that the "crash invariants" (ie, recovery
postconditions with only lower-level recovery) must always include the
postcondition. This can lead to some annoying duplication but is
otherwise fine.