Extract axiomatic prog
We extract prog to ReaderT Config IO, a monad with access to an
environment of type Config. In practice programs operate over two disks,
so Config comes from the Haskell implementation in TwoDiskEnvironment
(side note: Config could be renamed Environment). To plug in the two
disk operations, we first construct an Interface in Coq with axioms to
fill in (since constructing these dependently-typed records in Haskell
is more annoying), and then just plug in the read/write/diskSize axioms
via [Extract Constant].

This causes the extracted code to depend on hand-written code, so we
also have to fix up more imports. That implementation is becoming
brittle, since files are listed once in fiximports.py and again in the
Makefile to actually get fiximports.py called; we might want to call the
pre-processing script that adds the {-# OPTIONS_GHC -F -pgmF
./fiximports.py #-} pragma to every extracted file. We also might switch
to extracting a single monolithic file and only fixing one import list.