Implement server loop in Coq
Resulted in a fair bit of moving code around in extraction, but the new
implementation does work.

The server doesn't exit cleanly. The loop that processes requests is
infinite and ignores the commands, and blocks the connection thread from
ever exiting. The Coq loop is run in a forked thread and does understand
the disconnect command, but can't signal to stop processing commands.
Instead, it exits and then the server happily continues accepting
connections.

We should only allow a single simultaneous connection, terminate the
command processing upon disconnect (if nothing else, recognize the
disconnect command in Haskell), and exit the whole program after the
first connection terminates.