Support initializing with disks missing
Cleans up the environment handling code so that the state tracked for
each disk is opaque. Instead, there are just methods to get a Maybe Fd,
encapsulating failing early if the file never existed and checking if
the file exists before returning the persistent handle.

Handling the exception is a bit unwieldy, but it was the best way I
could see from the Control.Exception API to filter to a specific type of
exception.