Switch to a more physical async representation
For the physical disk use only the cached and durable values, while for
the abstraction in specs use a history of written values.

Many of the definitions have been significantly generalized to support
working with the two types of async disks. These includes defining
generic pointwise_rel and diskMap for disks and using a typeclass to
have common methods for working with async disk blocks.