Fix proof of rec_noop
Had to fix a really tricky inductive invariant issue: [rec_noop_compose]
formerly required a plain spec for the recovery procedure that had the
right precondition and implied the abstraction remained unchanged. This
spec had appropriate pre/recovery to be inductive and the proof showed
its idempotence.

The recovery condition used was too strong in practice, but it seemed
difficult to change to any other invariant. The solution now used is to
allow any idempotent spec for recovery and impose the appropriate
input-output behavior. Thus recovery can use its own fully general loop
invariant (that does not even imply invariant in the precondition) and
just show that specializing the precondition to the invariant, via
requiring an abstraction, results in the correct input-output behavior
(pre/post implication), irrespective of the resulting recovery
postcondition.