checkpoint exploration into proof-with-crashes without [prog_spec]
certainly seems doable but kind-of messy..

the thing that's bugging me is the duplication of specs between the
semantics of a layer (i.e., [op_step]) and the [prog_spec]-style
re-statement of the same thing.

so maybe the right question is, why not go all the way to [prog_spec],
and not bother with the [op_step] style semantics of a layer at all?
this is basically the FSCQ plan: no explicit layers, just hoare-style
specs everywhere with different levels of rep invariants..