unify movers and commutativity
turns out our earlier notion of commutativity was just "right-movers"..

i guess if everything is a right-mover, and there are no non-movers or
left-movers, that was fine too!  which is why Example.v worked fine.

in Example2.v, everything is a right-mover except for Release (which is
a left-mover), so we need to figure out some plan for that.