checkpoint an alternative to cofix for encoding infinite traces
instead of having real coinductive [exec], the [exec] is now indexed by
a maximum counter of execution steps.  [proc] is coinductive, but [exec]
and [trace] are always some finite prefix of an execution (an execution
might terminate early, if all threads finish, or it might stop because
the maximum counter expired).

this may allow us to keep regular induction instead of the cofix mess,
but still capture the safety properties we care about (any finite prefix
is safe).