fully prove and use release left-moverness
one trick: make [Release] always-enabled by adding nilpotent transitions
when Release is called but the lock is not held by the caller (i.e., Release
does nothing when called in that context).

next step is to prove, in ExampleProtocol.v, that this is a safe plan: i.e.,
if programs follow our locking syntactic rules, they will never call [Release]
in a way that triggers these nilpotent transitions.