prove that our weird hack with [Release] being always-enabled is OK
i.e., prove in ExampleProtocol.v that following the syntactic rules
for [Acquire] before [Release] is sufficient to never trigger these
extra step semantics for [Release].

the one issue to fix later is that we define [follows_protocol_proc]
as an [Inductive] instead of [CoInductive], so this theorem might not
apply.

luckily, it seems like we will eventually catch it when we try to use the
[protocol_ok] theorem from ExampleProtocol.v.  it will require proving
[follows_protocol_proc] for every thread's code, and if any thread's code
is possibly-infinite, the user will not be able to provide this premise.