simplify traces and events
no more low events at all.

high events are not call and return, but just opaque blobs.  no need to
specify an [opHiT] any more either, because there's no restriction on
what can be logged in a high event.

the point of the events, as far as i can tell, is that individual
opcodes do not have any events inside of them.  since the top-level
correctness condition is preserving events, this gives freedom to change
the implementation of an opcode (since no events are produced there),
as long as the eventual return value is the same (since that will cause
the continuation to produce the same events).

as an example of events, i suspect a mail server could log things like
"delivered message M for user U" and "user U downloaded message M".
the theorem then says, these will be the same across refinements.