checkpoint towards end-to-end link and proof in Example2.v
remaining issue: locking protocol and arbitrary starting states.

what if the code starts in a state where tid 5 has the lock held?
and tid 5's code happens to call [Acquire] right away?

this intuitively seems safe (because Acquire will block forever)
but the current protocol logic says this is illegal because a proc
cannot call [Acquire] if it's holding the lock.