get Example2.v to fully prove again
re-introduce the "always-enabled hack" but in a slightly more general way,
and without having to manually write the hacky step rules.

this works by separating out the protocol-restricted step relation from the
actual protocol restrictions, and by defining the protocol-restricted step
relation to be enabled (but with a noop effect) if the protocol restriction
doesn't hold.