fix up down to MailFSAPI
surprising complication in DeliverListTidAPI: ListTid is no longer a
right-mover in the presence of Delete, because Delete might remove a
file that ListTid was about to return!  fix: allow ListTid to return
extra stuff (i.e., allowed to return files that are missing, but every
existing file MUST be in ListTid's result set).