sketch of how to generate much cleaner code for the mail server
the result actually looks readable and kind-of similar to the Go code,
and is provably equivalent to the previous code (modulo one lemma about
Until loops).

currently this generates cleaner code for just 1 SMTP process, but should
be possible to extend this to something more general.

don't know how much this impacts performance yet.