Work out framework for abstracting store buffers
Invariant and abstraction combine really nicely - just need an
intermediate description of Inc and Dec in terms of low-level
operations. Not very satisfying for longer critical sections (the actual
sequence of reads and writes doesn't matter, and there should be a layer
that gives this fact), but it works for the example.