Switch to a new thread state representation
Use a function for convenient access, with a finiteness proof. This
representation allows gaps and has an upper bound for allocating a
thread and constructing thread states from fixpoints, like the list
representation. Unlike a list, it has a canonical representation (lists
can have extra NoProc elements at the end), so that if two thread states
are equal for all tids, they are Liebniz equal.