Progress on Mail conversion; having trouble with DeliverImpl proofs.
I can't get the Ltac to do quite as much as it used to, so there are
some additional subgoals in createwritetmp_right_mover that I don't
understand. Similar subgoals are also showing up towards the end of
unlinktmp_left_mover.

I got the conversion for MailServerComposedImpl to go through, but am
having type issues with some of the definitions in
MailServerLockAbsImpl.